Ahh... Would help if i staged changes...

+++ src/main/java/com/brandon3055/brandonscore/lib/datamanager/ManagedByte.java
@@ -12,8 +12,8 @@ public class ManagedByte extends AbstractManagedData {
     public byte value;
     private byte lastTickValue;
 
-    public ManagedByte(byte value) {
-        this.value = this.lastTickValue = value;
+    public ManagedByte(int value) {
+        this.value = this.lastTickValue = (byte) value;
     }
 
     @Override
